wip: fix segfault

+++ cxx/main.cpp
@@ -401,14 +401,14 @@ private:
             return;
         }
 
-        for (size_t i = 0; i < frame_data.size(); ++i)
+        for (int i = 0; i < frame_data.size(); ++i)
         {
             // reset all leds for next frame
             reset();
 
-            for (size_t j = frame_data[i].size()-1; j >= 0; --j)
+            for (int j = frame_data[i].size()-1; j >= 0; --j)
             {
-                for (size_t k = frame_data[i][j].size()-1; k >= 0; --k)
+                for (int k = frame_data[i][j].size()-1; k >= 0; --k)
                 {
                     const auto led_value = frame_data[i][j][k];
